fix a bug with song ordering
Previously, songs would not be correctly sorted until they were voted on for
the first time. They were always added to the end of the queue. If some
song already on the queue has a negative score, this ordering is
incorrect. Now, songs are added to their correct queue position.